scripts/requirements-media-kit.txt
fpdf2>=2.8.0
pillow>=11.0.0
